+++ packages/backend/src/nest/qss/qss.service.spec.ts
@@ -15,6 +15,7 @@ import {
   CommunityOperationStatus,
   CreateCommunityStatus,
   CommunitySignInMessage,
+  QSSDataSyncMessage,
 } from './qss.types'
 import { createLogger } from '../common/logger'
 import { Community, Identity } from '@quiet/types'
@@ -26,6 +27,22 @@ import { randomBytes } from 'crypto'
 import waitForExpect from 'wait-for-expect'
 import * as uint8arrays from 'uint8arrays'
 import { JoinStatus } from '../libp2p/libp2p.auth'
+import { OrbitDbService } from '../storage/orbitDb/orbitDb.service'
+import { Libp2pService } from '../libp2p/libp2p.service'
+import { IpfsService } from '../ipfs/ipfs.service'
+import { LocalDbService } from '../local-db/local-db.service'
+import { Libp2pNodeParams } from '../libp2p/libp2p.types'
+import { spawnLibp2pInstancesInMemory } from '../common/test-utils'
+import * as fs from 'fs'
+import { EventsType } from '@orbitdb/core'
+import { EventsWithStorage } from '../storage/orbitDb/eventsWithStorage'
+import { MessagesAccessController } from '../storage/channels/messages/orbitdb/MessagesAccessController'
+import { EncryptedAndSignedPayload, EncryptionScopeType } from '../auth/services/crypto/types'
+import { RoleName } from '../auth/services/roles/roles'
+import { IpfsFileManagerModule } from '../ipfs-file-manager/ipfs-file-manager.module'
+import { IpfsModule } from '../ipfs/ipfs.module'
+import { logEntryToLogUpdate } from '../storage/orbitDb/util'
+import { OrbitDbModule } from '../storage/orbitDb/orbitdb.module'
 
 describe('QSSService', () => {
   let store: Store
@@ -34,8 +51,14 @@ describe('QSSService', () => {
   let qssClient: QSSClient
   let qssService: QSSService
   let sigchainService: SigChainService
+  let libp2pService: Libp2pService
+  let ipfsService: IpfsService
+  let orbitDbService: OrbitDbService
+  let localDbService: LocalDbService
+  let libp2pParams: Libp2pNodeParams
   let mockedCreateSocket: any
   let mockedSendMessage: any
+  let addPendingMessageSpy: any
   let mockedAllowed: any
   let community: Community
   let userIdentity: Identity
@@ -50,10 +73,20 @@ describe('QSSService', () => {
     factory = await getReduxStoreFactory(store)
 
     module = await Test.createTestingModule({
-      imports: [TestModule, QSSModule, SigChainModule],
+      imports: [TestModule, SigChainModule, IpfsFileManagerModule, IpfsModule, OrbitDbModule, QSSModule],
     }).compile()
     qssService = module.get<QSSService>(QSSService)
     qssClient = module.get<QSSClient>(QSSClient)
+    libp2pService = await module.resolve(Libp2pService)
+    libp2pParams = (await spawnLibp2pInstancesInMemory([module]))[0]
+
+    ipfsService = await module.resolve(IpfsService)
+    await ipfsService.createInstance()
+
+    localDbService = await module.resolve(LocalDbService)
+    orbitDbService = await module.resolve(OrbitDbService)
+    await orbitDbService.create(ipfsService.ipfsInstance!)
+
     mockedCreateSocket = jest
       .spyOn(qssClient, 'createSocket')
       .mockImplementation(async (_qssEndpoint: string | undefined): Promise<ClientSocket> => {
@@ -80,11 +113,19 @@ describe('QSSService', () => {
   })
 
   afterEach(async () => {
-    await module.close()
+    await orbitDbService?.stop()
+    if (fs.existsSync(orbitDbService.orbitDbDir)) {
+      fs.rmSync(orbitDbService.orbitDbDir, { recursive: true })
+    }
+    await ipfsService?.stop()
+    await libp2pService?.close(true)
+    await localDbService?.close()
+    await module?.close()
     mockedCreateSocket.mockRestore()
     if (mockedSendMessage != null) {
       mockedSendMessage.mockRestore()
     }
+    addPendingMessageSpy = null
     if (mockedAllowed != null) {
       mockedAllowed.mockRestore()
     }
@@ -131,22 +172,18 @@ describe('QSSService', () => {
               case WebsocketEvents.GEN_PUB_KEYS:
                 return {
                   ts: DateTime.utc().toMillis(),
+                  status: CommunityOperationStatus.SUCCESS,
                   payload: {
-                    status: CommunityOperationStatus.SUCCESS,
-                    payload: {
-                      teamId: sigchainService.team?.id,
-                      keys: redactKeys(
-                        createKeyset({ type: 'SERVER', name: 'localhost' }, randomBytes(32).toString('base64'))
-                      ),
-                    },
+                    teamId: sigchainService.team?.id,
+                    keys: redactKeys(
+                      createKeyset({ type: 'SERVER', name: 'localhost' }, randomBytes(32).toString('base64'))
+                    ),
                   },
                 } as T
               case WebsocketEvents.CREATE_COMMUNITY:
                 return {
                   ts: DateTime.utc().toMillis(),
-                  payload: {
-                    status: CreateCommunityStatus.SUCCESS,
-                  },
+                  status: CreateCommunityStatus.SUCCESS,
                 } as T
               default:
                 return undefined
@@ -165,6 +202,7 @@ describe('QSSService', () => {
           WebsocketEvents.GEN_PUB_KEYS,
           expect.objectContaining({
             ts: expect.any(Number),
+            status: CommunityOperationStatus.SENDING,
             payload: {
               teamId: sigchainService.team.id,
             },
@@ -209,17 +247,13 @@ describe('QSSService', () => {
               case WebsocketEvents.GEN_PUB_KEYS:
                 return {
                   ts: DateTime.utc().toMillis(),
-                  payload: {
-                    status: CommunityOperationStatus.ERROR,
-                    reason: 'Failed to create server keys',
-                  },
+                  status: CommunityOperationStatus.ERROR,
+                  reason: 'Failed to create server keys',
                 } as T
               case WebsocketEvents.CREATE_COMMUNITY:
                 return {
                   ts: DateTime.utc().toMillis(),
-                  payload: {
-                    status: CreateCommunityStatus.SUCCESS,
-                  },
+                  status: CreateCommunityStatus.SUCCESS,
                 } as T
               default:
                 return undefined
@@ -238,6 +272,7 @@ describe('QSSService', () => {
           WebsocketEvents.GEN_PUB_KEYS,
           expect.objectContaining({
             ts: expect.any(Number),
+            status: CommunityOperationStatus.SENDING,
             payload: {
               teamId: sigchainService.team.id,
             },
@@ -260,23 +295,19 @@ describe('QSSService', () => {
               case WebsocketEvents.GEN_PUB_KEYS:
                 return {
                   ts: DateTime.utc().toMillis(),
+                  status: CommunityOperationStatus.SUCCESS,
                   payload: {
-                    status: CommunityOperationStatus.SUCCESS,
-                    payload: {
-                      teamId: sigchainService.team?.id,
-                      keys: redactKeys(
-                        createKeyset({ type: 'SERVER', name: 'localhost' }, randomBytes(32).toString('base64'))
-                      ),
-                    },
+                    teamId: sigchainService.team?.id,
+                    keys: redactKeys(
+                      createKeyset({ type: 'SERVER', name: 'localhost' }, randomBytes(32).toString('base64'))
+                    ),
                   },
                 } as T
               case WebsocketEvents.CREATE_COMMUNITY:
                 return {
                   ts: DateTime.utc().toMillis(),
-                  payload: {
-                    status: CreateCommunityStatus.ERROR,
-                    reason: 'Failed to create community',
-                  },
+                  status: CreateCommunityStatus.ERROR,
+                  reason: 'Failed to create community',
                 } as T
               default:
                 return undefined
@@ -294,6 +325,7 @@ describe('QSSService', () => {
           WebsocketEvents.GEN_PUB_KEYS,
           expect.objectContaining({
             ts: expect.any(Number),
+            status: CommunityOperationStatus.SENDING,
             payload: {
               teamId: sigchainService.team.id,
             },
@@ -351,9 +383,7 @@ describe('QSSService', () => {
               case WebsocketEvents.CREATE_COMMUNITY:
                 return {
                   ts: DateTime.utc().toMillis(),
-                  payload: {
-                    status: CreateCommunityStatus.SUCCESS,
-                  },
+                  status: CreateCommunityStatus.SUCCESS,
                 } as T
               default:
                 return undefined
@@ -382,9 +412,7 @@ describe('QSSService', () => {
               case WebsocketEvents.SIGN_IN_COMMUNITY:
                 return {
                   ts: DateTime.utc().toMillis(),
-                  payload: {
-                    status: CommunityOperationStatus.SUCCESS,
-                  },
+                  status: CommunityOperationStatus.SUCCESS,
                 } as T
               default:
                 return undefined
@@ -404,12 +432,10 @@ describe('QSSService', () => {
           WebsocketEvents.SIGN_IN_COMMUNITY,
           expect.objectContaining({
             ts: expect.any(Number),
+            status: CommunityOperationStatus.SUCCESS,
             payload: {
-              status: CommunityOperationStatus.SUCCESS,
-              payload: {
-                userId: sigchainService.user.userId,
-                teamId: sigchainService.team.id,
-              },
+              userId: sigchainService.user.userId,
+              teamId: sigchainService.team.id,
             },
           } as CommunitySignInMessage),
           true
@@ -431,9 +457,7 @@ describe('QSSService', () => {
               case WebsocketEvents.SIGN_IN_COMMUNITY:
                 return {
                   ts: DateTime.utc().toMillis(),
-                  payload: {
-                    status: CommunityOperationStatus.SUCCESS,
-                  },
+                  status: CommunityOperationStatus.SUCCESS,
                 } as T
               default:
                 return undefined
@@ -456,10 +480,8 @@ describe('QSSService', () => {
               case WebsocketEvents.SIGN_IN_COMMUNITY:
                 return {
                   ts: DateTime.utc().toMillis(),
-                  payload: {
-                    status: CommunityOperationStatus.ERROR,
-                    reason: 'Failed to sign in',
-                  },
+                  status: CommunityOperationStatus.ERROR,
+                  reason: 'Failed to sign in',
                 } as T
               default:
                 return undefined
@@ -484,12 +506,10 @@ describe('QSSService', () => {
           WebsocketEvents.SIGN_IN_COMMUNITY,
           expect.objectContaining({
             ts: expect.any(Number),
+            status: CommunityOperationStatus.SUCCESS,
             payload: {
-              status: CommunityOperationStatus.SUCCESS,
-              payload: {
-                userId: sigchainService.user.userId,
-                teamId: sigchainService.team.id,
-              },
+              userId: sigchainService.user.userId,
+              teamId: sigchainService.team.id,
             },
           } as CommunitySignInMessage),
           true
@@ -499,6 +519,134 @@ describe('QSSService', () => {
       expect(error?.message.includes('Error while signing in to community')).toBeTruthy()
       expect(qssService.joinStatus(sigchainService.team.id)).toBe(JoinStatus.NOT_STARTED)
       expect(mockedSendMessage).toHaveBeenCalledTimes(1)
+    })
+  })
+
+  describe('sendDataSyncMessage', () => {
+    it(`sends a successful data sync to QSS`, async () => {
+      mockedSendMessage = jest
+        .spyOn(qssClient, 'sendMessage')
+        .mockImplementation(
+          async <T>(event: WebsocketEvents, payload: unknown, withAck = false): Promise<T | undefined> => {
+            logger.debug('Sending event to QSS', event, payload, withAck)
+            if (!withAck) {
+              return undefined
+            }
+            switch (event) {
+              case WebsocketEvents.DATA_SYNC:
+                const { teamId, hash, hashedDbId } = (payload as QSSDataSyncMessage).payload!
+                return {
+                  ts: DateTime.utc().toMillis(),
+                  status: CommunityOperationStatus.SUCCESS,
+                  payload: {
+                    teamId,
+                    hash,
+                    hashedDbId,
+                  },
+                } as T
+              default:
+                return undefined
+            }
+          }
+        )
+      mockedAllowed = jest.spyOn(qssService, 'qssAllowed', 'get').mockReturnValue(true)
+      await qssService.connect('ws://localhost:3000')
+      expect(qssService.connected).toBeTruthy()
+
+      const db = await orbitDbService.open<EventsType<EncryptedAndSignedPayload>>(`channels.foobar`, {
+        type: 'events',
+        Database: EventsWithStorage(),
+        AccessController: MessagesAccessController({ write: ['*'] }),
+        sync: true,
+      })
+      const hash = await db.add(
+        sigchainService.activeChain.crypto.encryptAndSign('random message', {
+          type: EncryptionScopeType.ROLE,
+          name: RoleName.MEMBER,
+        })
+      )
+      const entry = await db.log.get(hash)
+      const update = logEntryToLogUpdate(entry, db.address)
+      const result = await qssService.sendDataSyncMessage(update)
+      await waitForExpect(() => {
+        expect(mockedSendMessage).toHaveBeenNthCalledWith(
+          1,
+          WebsocketEvents.DATA_SYNC,
+          expect.objectContaining({
+            ts: expect.any(Number),
+            status: CommunityOperationStatus.SENDING,
+            payload: {
+              teamId: sigchainService.team.id,
+              hash,
+              hashedDbId: expect.any(String),
+              encEntry: expect.any(Object),
+            },
+          } as QSSDataSyncMessage),
+          true
+        )
+      })
+      expect(result).toBe(true)
+      expect(mockedSendMessage).toHaveBeenCalledTimes(1)
+
+      const pendingMessages = await localDbService.getPendingQssSyncMessages()
+      expect(pendingMessages).toEqual({})
+    })
+
+    it(`fails to send data sync to QSS and writes pending message to local DB`, async () => {
+      mockedSendMessage = jest
+        .spyOn(qssClient, 'sendMessage')
+        .mockImplementation(
+          async <T>(event: WebsocketEvents, payload: unknown, withAck = false): Promise<T | undefined> => {
+            logger.debug('Sending event to QSS', event, payload, withAck)
+            if (!withAck) {
+              return undefined
+            }
+            switch (event) {
+              case WebsocketEvents.DATA_SYNC:
+                const { teamId, hash, hashedDbId } = (payload as QSSDataSyncMessage).payload!
+                return {
+                  ts: DateTime.utc().toMillis(),
+                  status: CommunityOperationStatus.SUCCESS,
+                  payload: {
+                    teamId,
+                    hash,
+                    hashedDbId,
+                  },
+                } as T
+              default:
+                return undefined
+            }
+          }
+        )
+      addPendingMessageSpy = jest.spyOn(localDbService, 'addPendingQssSyncMessage')
+      mockedCreateSocket.mockRestore()
+      mockedAllowed = jest.spyOn(qssService, 'qssAllowed', 'get').mockReturnValue(true)
+      await qssService.connect('ws://localhost:3000')
+      expect(qssService.connected).toBeFalsy()
+
+      const db = await orbitDbService.open<EventsType<EncryptedAndSignedPayload>>(`channels.foobar`, {
+        type: 'events',
+        Database: EventsWithStorage(),
+        AccessController: MessagesAccessController({ write: ['*'] }),
+        sync: true,
+      })
+      const hash = await db.add(
+        sigchainService.activeChain.crypto.encryptAndSign('random message', {
+          type: EncryptionScopeType.ROLE,
+          name: RoleName.MEMBER,
+        })
+      )
+      const entry = await db.log.get(hash)
+      const update = logEntryToLogUpdate(entry, db.address)
+      const result = await qssService.sendDataSyncMessage(update)
+      expect(result).toBe(undefined)
+      await waitForExpect(async () => {
+        expect(addPendingMessageSpy).toHaveBeenCalledTimes(1)
+      })
+      expect(mockedSendMessage).toHaveBeenCalledTimes(0)
+
+      const pendingMessages = await localDbService.getPendingQssSyncMessages()
+      expect(pendingMessages[db.address].length).toBe(1)
     })
   })
 })