+++ packages/desktop/src/main/main.ts
@@ -28,7 +28,11 @@ let SOCKET_IO_SECRET: string | undefined = undefined
 const updaterInterval = 15 * 60_000
 
 export const isDev = process.env.NODE_ENV === 'development'
-export const isE2Etest = process.env.E2E_TEST === 'true'
+export const isE2Etest = process.env.IS_E2E === 'true'
+
+if (isE2Etest) {
+  autoUpdater.autoInstallOnAppQuit = false
+}
 
 const webcrypto = new Crypto()